+++ jdaviz/configs/cubeviz/plugins/spectral_extraction/tests/test_spectral_extraction.py
@@ -85,7 +85,7 @@ def test_gauss_smooth_before_spec_extract(cubeviz_helper, spectrum1d_cube_with_u
     extract_plugin.function = "Sum"
     expected_uncert = 2
 
-    extract_plugin.spatial_subset = 'Subset 1'
+    extract_plugin.aperture = 'Subset 1'
     collapsed_spec = extract_plugin.collapse_to_spectrum()
 
     # this single pixel has two wavelengths, and all uncertainties are unity
@@ -94,7 +94,7 @@ def test_gauss_smooth_before_spec_extract(cubeviz_helper, spectrum1d_cube_with_u
     assert np.all(np.equal(collapsed_spec.uncertainty.array, 1))
 
     # this two-pixel region has four unmasked data points per wavelength:
-    extract_plugin.spatial_subset = 'Subset 2'
+    extract_plugin.aperture = 'Subset 2'
     collapsed_spec_2 = extract_plugin.collapse_to_spectrum()
     assert np.all(np.equal(collapsed_spec_2.uncertainty.array, expected_uncert))
 
@@ -129,7 +129,7 @@ def test_subset(
     plg.function = function
 
     # single pixel region:
-    plg.spatial_subset = 'Subset 1'
+    plg.aperture = 'Subset 1'
     collapsed_spec_1 = plg.collapse_to_spectrum()
 
     # this single pixel has two wavelengths, and all uncertainties are unity
@@ -138,7 +138,7 @@ def test_subset(
     assert np.all(np.equal(collapsed_spec_1.uncertainty.array, 1))
 
     # this two-pixel region has four unmasked data points per wavelength:
-    plg.spatial_subset = 'Subset 2'
+    plg.aperture = 'Subset 2'
     collapsed_spec_2 = plg.collapse_to_spectrum()
 
     assert np.all(np.equal(collapsed_spec_2.uncertainty.array, expected_uncert))
@@ -192,3 +192,42 @@ def test_save_collapsed_to_fits(cubeviz_helper, spectrum1d_cube_with_uncerts, tm
     with pytest.raises(ValueError, match="Invalid path=/this/path/doesnt"):
         extract_plugin._obj.vue_save_as_fits()
     extract_plugin._obj.filename == fname  # set back to original filename
+
+
+def test_aperture_markers(cubeviz_helper, spectrum1d_cube):
+
+    cubeviz_helper.load_data(spectrum1d_cube)
+    cubeviz_helper.load_regions([CirclePixelRegion(PixCoord(0.5, 0), radius=1.2)])
+
+    extract_plg = cubeviz_helper.plugins['Spectral Extraction']
+    slice_plg = cubeviz_helper.plugins['Slice']
+
+    mark = extract_plg.aperture.marks[0]
+    assert not mark.visible
+    with extract_plg.as_active():
+        assert mark.visible
+        assert not len(mark.x)
+
+        extract_plg.aperture = 'Subset 1'
+        before_x = mark.x
+        assert len(before_x) > 0
+
+        # sample cube only has 2 slices with wavelengths [4.62280007e-07 4.62360028e-07] m
+        slice_plg.slice = 1
+        assert mark.x[1] == before_x[1]
+
+        slice_plg.slice = 0
+        extract_plg._obj.dev_cone_support = True
+        extract_plg._obj.wavelength_dependent = True
+        assert mark.x[1] == before_x[1]
+
+        slice_plg.slice = 1
+        assert mark.x[1] != before_x[1]
+
+        extract_plg._obj.vue_goto_reference_wavelength()
+        assert slice_plg.slice == 0
+
+        slice_plg.slice = 1
+        extract_plg._obj.vue_adopt_slice_as_reference()
+        extract_plg._obj.vue_goto_reference_wavelength()
+        assert slice_plg.slice == 1